requirements.txt
PyMuPDF
streamlit
pip install google-generativeai python-dotenv
google-generativeai
python-dotenv
duckduckgo-search
pip install easyocr opencv-python-headless Pillow
